Update patch set 1
Patch Set 1:

I tried multiple DSP configurations and every kernel option I could think of to fix the issue. Eventually we tracked it down to the updated spi_aic3254.c and once it was reverted things worked just fine.

I'm still looking for time to break down the newer source and try to find the exact source of the error, but for now this is a non-obtrusive workaround which only touches the Saga handsets.

Patch-set: 1