Update patch set 1
Patch Set 1:

Yes, I realize what panel type means.

My point is that probably there are both vivo and vivow devices with this type of panel -> thus the if you mention is reached for both vivo and vivow phones.
The reason I have an ifndef instead of just removing the whole if is that I did not want to affect the vivow devices - I don't know how they will behave and have no means for testing.

If you consider it is safe then ofcourse you can entirely remove the whole if.

Patch-set: 1